[TTS] Separate TTS tokenization and g2p util to fix circular import
