+++ nemo/collections/tts/g2p/modules.py
@@ -25,15 +25,13 @@ import nltk
 import torch
 
 from nemo.collections.common.tokenizers.text_to_speech.ipa_lexicon import validate_locale
-from nemo.collections.tts.g2p.data.data_utils import (
-    GRAPHEME_CASE_MIXED,
-    GRAPHEME_CASE_UPPER,
+from nemo.collections.common.tokenizers.text_to_speech.tokenizer_utils import (
     LATIN_CHARS_ALL,
     any_locale_word_tokenize,
     english_word_tokenize,
     normalize_unicode_text,
-    set_grapheme_case,
 )
+from nemo.collections.tts.g2p.data.data_utils import GRAPHEME_CASE_MIXED, GRAPHEME_CASE_UPPER, set_grapheme_case
 from nemo.utils import logging
 from nemo.utils.decorators import experimental
 from nemo.utils.get_rank import is_global_rank_zero
